cleaning up project structure, blog and project folders created components moved
